split files in single files for start, stop,restart
